refactor: data/news image interface
