+++ src/components/templates/NewsSlider.tsx
@@ -52,10 +52,10 @@ export const NewsSlider = ({ posts }: NewsSliderProps) => {
           href='/newsroom'
           className='group inline-flex items-center'
         >
-          <Title size='five' margin={false} className='mr-4 hover:underline'>
+          <Title size='five' margin={false} className='mr-2 hover:underline'>
             {t('content.newsSlider.readMore')}
           </Title>
-          <span className='relative top-0.5 block -rotate-45 transform transition duration-100 group-hover:rotate-0 group-hover:text-primary-500'>
+          <span className='relative block -rotate-45 transform transition duration-100 group-hover:rotate-0 group-hover:text-primary-500'>
             <VscArrowRight size={24} />
           </span>
         </UnderlineLink>
