+++ src/interfaces/Image.ts
@@ -1,0 +1,10 @@
+export interface Image {
+  data: ImageAttribute;
+}
+
+interface ImageAttribute {
+  url: string;
+  width: number;
+  height: number;
+  name: string;
+}
